pyproject.toml
[project]
name = "crawl4ai-mcp"
version = "0.2.0"
description = "MCP server for integrating web crawling and RAG into AI agents and AI coding assistants"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "crawl4ai>=0.8.6",
    "fastmcp>=2.0.0",
    "sqlmodel>=0.0.14",
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.2.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "openai>=1.71.0",
    "numpy>=1.26.0",
    "nltk>=3.8.0",
    "sentence-transformers>=3.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-mock>=3.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
pythonpath = ["."]
markers = [
    "unit: fast, offline tests using mocks",
    "integration: tests requiring live services",
]
addopts = "-m 'not integration'"

[tool.coverage.run]
source = ["src"]
omit = [
    "src/__init__.py",
    "src/crawl4ai_mcp.py",
    "src/reembed_documents.py",
]

[tool.coverage.report]
fail_under = 100
show_missing = true
